trainer/trainer.py: Fix train epoch logic and saving updated variational parameters
base/base_data_loader.py: Initialise optimizer_phi within the trainer
class
data_loader/data_loaders.py: Use pinned memory and re-use the identity
grid

model/metric.py: Fix bug in the SSD metric
train.py: Initialise optimiser_phi within the trainer class

base/base_trainer.py: Initialise optimiser_phi within the trainer class
trainer/trainer.py: Fix train epoch logic and saving updated variational
parameters

utils/util.py: Code clean-up

+++ utils/util.py
@@ -65,9 +65,9 @@ def init_identity_grid(im_dim):
     ny = sz[1]
     nx = sz[2]
 
-    x = torch.linspace(-1, 1, steps=nx).to(dtype=torch.float32)
-    y = torch.linspace(-1, 1, steps=ny).to(dtype=torch.float32)
-    z = torch.linspace(-1, 1, steps=nz).to(dtype=torch.float32)
+    x = torch.linspace(-1, 1, steps=nx)
+    y = torch.linspace(-1, 1, steps=ny)
+    z = torch.linspace(-1, 1, steps=nz)
 
     x = x.expand(ny, -1).expand(nz, -1, -1)
     y = y.expand(nx, -1).expand(nz, -1, -1).transpose(1, 2)
